Added exercises 2.16 and 2.17 from Part 2e

+++ part2/phonebook/src/App.jsx
@@ -3,6 +3,7 @@ import personService from './services/persons';
 import Search from './components/Search';
 import AddForm from './components/AddForm';
 import Numbers from './components/Numbers';
+import Notification from './components/Notification';
 
 
 const App = () => {
@@ -10,6 +11,8 @@ const App = () => {
   const [newName, setNewName] = useState('');
   const [newNumber, setNewNumber] = useState('');
   const [search, setSearch] = useState('');
+  const [notification, setNotification] = useState('');
+  const [messageType, setMessageType] = useState('');
 
   useEffect(() => {
     const fetchData = async () => {
@@ -35,20 +38,37 @@ const App = () => {
     setSearch(event.target.value.toLowerCase().trimStart());
   };
 
-  const handleNumberUpdate = async (person) => {
+  const handleNumberUpdate = async (personId, updatedNumber) => {
+    const personToUpdate = persons.find(p => p.id === personId);
+    if (!personToUpdate) return;
+
     try {
-      const returnedPerson = await personService.updateNumber(person.id, { ...person, number: newNumber.trim() });
-      setPersons(persons.map((person) => (person.id !== returnedPerson.id ? person : returnedPerson)));
+      const updatedPerson = { ...personToUpdate, number: updatedNumber.trim() };
+      const returnedPerson = await personService.updateNumber(personId, updatedPerson);
+      setPersons(persons.map(p => (p.id !== returnedPerson.id ? p : returnedPerson)));
+      handleNotification('success-message', `${returnedPerson.name}'s number has been updated!`);
     } catch (err) {
       console.error(err);
+
+      handleNotification('error-message', `${personToUpdate.name}'s has already been removed from server`);
+      setPersons(persons.filter(p => p.id !== personId));
     }
+  }
+
+  const handleNotification = (type, message) => {
+    setMessageType(type);
+    setNotification(message);
+    setTimeout(() => {
+      setMessageType('');
+      setNotification('');
+    }, 5000);
   }
 
   // Checks for invalid input, such as empty fields
   const verifyData = () => {
     // Verifies if both fields have been filled
     if (!newName.trim() || !newNumber.trim()) {
-      alert('All fields are required');
+      handleNotification('error-message', 'All fields are required');
       return false;
     }
     return true;
@@ -59,7 +79,7 @@ const App = () => {
       if (person.name.toLowerCase() === newName.toLowerCase().trim()) {
         const confirm = window.confirm(`${newName.trim()} is already added to the Phonebook, replace the old number with a new one?`);
         if (confirm) {
-          await handleNumberUpdate(person);
+          await handleNumberUpdate(person.id, newNumber.trim());
         }
         return true;
       }
@@ -73,12 +93,14 @@ const App = () => {
     // If the data verified is valid, proceed to add a new person
     if (verifyData()) {
       try {
-        if (!checkDuplicates()) {
+        const isDuplicate = await checkDuplicates();
+        if (!isDuplicate) {
           const newPerson = { name: newName.trim(), number: newNumber.trim() }
           const savedPerson = await personService.storeData(newPerson);
           setPersons(persons.concat(savedPerson));
           setNewName('');
           setNewNumber('');
+          handleNotification('success-message', `${savedPerson.name} was added to the Phonebook`);
         }
       } catch (err) {
         console.error(err);
@@ -86,12 +108,32 @@ const App = () => {
     }
   };
 
+  // Checks if the ID is still present on the server database before attempting to remove it
+  const verifyId = async (id) => {
+    try {
+      const person = await personService.getDataById(id);
+      if (!person) {
+        return false;
+      }
+      return person;
+    } catch (err) {
+      console.error(err);
+    }
+  }
+
   const handleDelete = async (id, name) => {
     try {
+      const personToRemove = await verifyId(id);
+      if (!personToRemove) {
+        handleNotification('error-message', `${name} was already removed from the server`);
+        setPersons(persons.filter(p => p.id !== id));
+        return;
+      }
       const confirm = window.confirm(`Remove ${name} from the Phonebook?`);
       if (confirm) {
         await personService.removeData(id);
         setPersons(persons.filter(p => p.id !== id));
+        handleNotification('success-message', `${name} was removed from the Phonebook`);
       }
     } catch (err) {
       console.error('Error removing person: ', err);
@@ -101,6 +143,10 @@ const App = () => {
   return (
     <div>
       <h1>Phonebook</h1>
+      <Notification
+        messageType={messageType}
+        notification={notification}
+      />
       <Search
         search={search}
         handleSearch={handleSearch}
